# Faster parallel array partition and choose