--- a/packages.config
+++ b/packages.config
@@ -18,4 +18,6 @@
   <package id="Microsoft.VisualFSharp.Type.Providers.Redist" version="1.0.0" />
   <package id="Microsoft.VisualFSharp.Msbuild.15.0" version="1.0.1" />
   <package id="System.ValueTuple" version="4.0.0-rc3-24212-01" />
+  <package id="BenchmarkDotNet" version="0.9.8"/>
+  <package id="BenchmarkDotNet.Diagnostics.Windows" version="0.9.8"/>
 </packages>