--- a/tests/scripts/array-perf/array-perf.fs
+++ b/tests/scripts/array-perf/array-perf.fs
@@ -168,7 +168,7 @@ type ArrayBenchmark () =
 
     [<Benchmark>]
     member self.New () =
-        array |> partition (fun x -> x % 2 = 0)
+        array |> partitionNew (fun x -> x % 2 = 0)
         
     //Create any number of benchmarks