Runtime is about 1.3x faster on my machine (4 core i7 mobile), with no memory use penalty.  Would love input on this one, there may be some ways to improve on the approach. I am deep in uncharted Parallel.For territory.

BenchmarkDotNet results, partitioning a random int array with `(fun x -> x % 2 = 0)`

Note that the benchmark results are somewhat stochastic, they will bounce around from 35% to 10% better each time I run the suite, there doesn't appear to be a general trend where it gets better or worse moving from 1 million to 10 million.  I've also used more complex predicates and the relative differences are similar.

```ini

Host Process Environment Information:
BenchmarkDotNet=v0.9.8.0
OS=Microsoft Windows NT 6.2.9200.0
Processor=Intel(R) Core(TM) i7-4712HQ CPU 2.30GHz, ProcessorCount=8
Frequency=2240907 ticks, Resolution=446.2479 ns, Timer=TSC
CLR=MS.NET 4.0.30319.42000, Arch=64-bit RELEASE [RyuJIT]
GC=Concurrent Workstation
JitModules=clrjit-v4.6.1590.0

Type=SIMDBenchmark  Mode=Throughput  Platform=X64  
Jit=RyuJit  GarbageCollection=Concurrent Workstation  

```

### Partition

   Method |   Length |          Median |        StdDev | Scaled | Gen 0 | Gen 1 | Gen 2 | Bytes Allocated/Op |
--------- |--------- |---------------- |-------------- |------- |------ |------ |------ |------------------- |
 **Original** |     **1000** |      **21.8514 us** |     **0.5300 us** |   **1.00** |  **0.16** |     **-** |     **-** |           **3,471.77** |
      New |     1000 |      20.5297 us |     0.8840 us |   0.94 |  0.17 |     - |     - |           3,489.75 |
 **Original** |    **10000** |     **160.0466 us** |     **3.1249 us** |   **1.00** |  **1.21** |     **-** |     **-** |          **28,955.03** |
      New |    10000 |     118.1885 us |     2.8572 us |   0.74 |  1.20 |     - |     - |          28,666.02 |
 **Original** |   **100000** |   **1,282.9827 us** |     **7.3705 us** |   **1.00** |     **-** |     **-** | **10.17** |         **211,334.08** |
      New |   100000 |     917.0063 us |    17.4501 us |   0.71 |     - |     - |  7.27 |         151,441.53 |
 **Original** |  **1000000** |  **12,467.9427 us** |   **728.8799 us** |   **1.00** |     **-** |     **-** | **65.99** |       **2,353,833.73** |
      New |  1000000 |   9,700.7108 us |   990.4339 us |   0.78 |     - |     - | 65.24 |       2,309,151.64 |
   Scalar |  1000000 |  13,680.7468 us |   212.0151 us |   1.10 |  3.25 |     - | 37.57 |       4,361,586.45 |
 **Original** | **10000000** | **125,043.1745 us** | **1,753.3497 us** |   **1.00** |     **-** |     **-** | **35.28** |      **29,670,713.02** |
      New | 10000000 |  86,908.7271 us | 1,472.4448 us |   0.70 |     - |     - | 35.53 |      29,909,345.33 |


### Choose

   Method |   Length |          Median |         StdDev | Scaled |    Gen 0 | Gen 1 | Gen 2 | Bytes Allocated/Op |
--------- |--------- |---------------- |--------------- |------- |--------- |------ |------ |------------------- |
 **Original** |     **1000** |      **24.1173 us** |      **1.8355 us** |   **1.00** |     **0.52** |     **-** |     **-** |          **13,721.83** |
      New |     1000 |      25.8602 us |      0.3443 us |   1.07 |     0.47 |     - |     - |          12,292.06 |
 **Original** |    **10000** |     **235.2287 us** |      **4.9615 us** |   **1.00** |     **4.55** |     **-** |     **-** |         **157,065.29** |
      New |    10000 |     190.6397 us |      5.1752 us |   0.81 |     4.72 |     - |     - |         167,680.01 |
 **Original** |   **100000** |   **2,352.3452 us** |     **58.5964 us** |   **1.00** |    **18.99** |     **-** | **10.23** |       **1,206,016.17** |
      New |   100000 |   1,696.8924 us |     38.6374 us |   0.72 |    18.30 |     - |  9.54 |       1,093,127.52 |
 **Original** |  **1000000** |  **23,316.6749 us** |    **899.0359 us** |   **1.00** |   **269.10** |  **1.62** | **41.29** |      **14,659,078.49** |
      New |  1000000 |  15,062.9299 us |  1,155.5808 us |   0.65 |   294.00 |  1.39 | 46.05 |      15,348,631.44 |
 **Original** | **10000000** | **178,700.9122 us** | **24,916.3155 us** |   **1.00** | **2,524.00** |     **-** | **38.00** |     **142,952,816.14** |
      New | 10000000 | 163,852.8507 us | 17,740.5511 us |   0.92 | 2,634.24 |     - | 35.28 |     141,815,184.40 |