@jackmott 2016-08-12
Whats does this mean, from the QA tests
"Conformance\DeclarationElements\CustomAttributes\AttributeUsage (AssemblyVersion03.fs) -- failed"