@jackmott 2016-08-12
Going to probably pull in the same trick for Parallel.choose in a minute.