@ReedCopsey 2016-08-12
Note that using Parallel.For can have bad repercussions in a situation where the algorithm is being used in a threaded context.  If this is called on a thread while the other threads are already saturated via CPU usage, best case, it'd be the same as doing it in serial, and worst (and more likely) case, significantly slower.

I'd be very careful about using Parallel loops within core library functions, since you can't control where they're going to be used, and this could be disastrous when used in highly threaded code.  It's also unexpected - you'd expect to be safe using core library functions when threading at a "higher level" in your application, and this would make that assumption invalid.