@jackmott 2016-08-12
It was a Parallel.For originally, I just pulled the accumulation into it, rather than iterating over the array again to count it up afterwards.

Also this is in a separate sub module - Array.Parallel,  which was already in place, so chance of accidental improper usage is low.