@cartermp 2016-08-12
@jackmott This is fantastic, as with your other PRs thus far.  👍 

Could you also add your test script to this PR and further PRs?  @smoothdeveloper and I were chatting, and he had mentioned that it would be great to have the perf test scripts for these changes in the repo so we can use them as a baseline for further improvements.

The location for this would be in `tests/scripts/{some-folder-name-which-makes-sense}`.

Thanks!