@jackmott 2016-08-12
@cartermp do you mean a script to test that performance has not regressed? Or for correctness?  The former, I'm not sure how to do, since the old function wouldn't exist any more.  The latter I assume already exists, since I haven't added a new function. Happy to help, need guidance. Thanks!