@KevinRansom 2016-08-12
Not a test case, more an indication of how the numbers were gathered.

It gives other developers an idea about how to go about testing their performance improvements.

Kevin

From: Jack Mott [mailto:notifications@github.com]
Sent: Friday, August 12, 2016 1:54 PM
To: Microsoft/visualfsharp <visualfsharp@noreply.github.com>
Subject: Re: [Microsoft/visualfsharp] Faster parallel array partition and choose (#1423)


@cartermp<https://github.com/cartermp> do you mean a script to test that performance has not regressed? Or for correctness? The former, I'm not sure how to do, since the old function wouldn't exist any more. The latter I assume already exists, since I haven't added a new function. Happy to help, need guidance. Thanks!

—
You are receiving this because you are subscribed to this thread.
Reply to this email directly, view it on GitHub<https://github.com/Microsoft/visualfsharp/pull/1423#issuecomment-239557184>, or mute the thread<https://github.com/notifications/unsubscribe-auth/AE76FqMB2ilp5Rnv74NQ5qGXi2MFhG0Rks5qfN19gaJpZM4Jisxr>.