@jackmott 2016-08-12
Gotcha.  Would it be ok to pull BenchmarkDotNet into the solution?  Are there other perf tests scripts I can use as a guide?