@ReedCopsey 2016-08-12
@jackmott Oh, great - sorry - just saw "for Array.choose" and wasn't thinking this was for Array.Parallel.  Sorry about that!  This is a great improvement - using the accumulators is definitely better.