@cartermp 2016-08-12
@jackmott I'd be fine with a package reference.  I think it would be best if the `.dll` for it wasn't checked in, though.  As for examples, it looks like all we have is [Compiler Perf Scripts](https://github.com/Microsoft/visualfsharp/tree/master/tests/scripts).  Not too much to go off of, so I think it's safe to say that you could do whatever you feel is reasonable and it'll be fine.

Biggest thing is just to ensure that anyone can clone the repo and begin running the scripts to see performance results themselves.  I'm also happy to iterate here and just get some scripts checked in while we figure out a good solution.