@KevinRansom 2016-08-12
Given that the tests will run on different hardware, and that there is a before and after component.  Seeing the results is less interesting than seeing the methodology.  And the ability to examine the test code to look for where it may impzact the result demonstrated.

Kevin

From: Phillip Carter [mailto:notifications@github.com]
Sent: Friday, August 12, 2016 2:44 PM
To: Microsoft/visualfsharp <visualfsharp@noreply.github.com>
Cc: Kevin Ransom <Kevin.Ransom@microsoft.com>; Comment <comment@noreply.github.com>
Subject: Re: [Microsoft/visualfsharp] Faster parallel array partition and choose (#1423)


@jackmott<https://github.com/jackmott> I'd be fine with a package reference. I think it would be best if the .dll for it wasn't checked in, though. As for examples, it looks like all we have is Compiler Perf Scripts<https://github.com/Microsoft/visualfsharp/tree/master/tests/scripts>. Not too much to go off of, so I think it's safe to say that you could do whatever you feel is reasonable and it'll be fine.

Biggest thing is just to ensure that anyone can clone the repo and begin running the scripts to see performance results themselves. I'm also happy to iterate here and just get some scripts checked in while we figure out a good solution.

—
You are receiving this because you commented.
Reply to this email directly, view it on GitHub<https://github.com/Microsoft/visualfsharp/pull/1423#issuecomment-239567197>, or mute the thread<https://github.com/notifications/unsubscribe-auth/AE76Fogf49IRykHqGp_rD9nUCCznM-nWks5qfOklgaJpZM4Jisxr>.