@jackmott 2016-08-13
Sample perf test script added. Let me know if I should rename/relocate/revise.
It is not formatted as an fsx, because I never use the REPL so am not familiar with it, but I could learn!