@jackmott 2016-08-13
It would be neat if you had coverage of any perf relevant functions, and maybe weekly, check today's master against 1 week ago, report any that have  regressed by some % for review.  Massively time consuming test, couldn't do it on every PR/commit I don't think.