@KevinRansom 2016-08-13
@dotnet-bot test this please