@KevinRansom 2016-08-13
Filure is due to app veyor timeout, it will be fixed by: https://github.com/Microsoft/visualfsharp/pull/1425