@veikkoeeva 2016-08-14
> It would be neat if you had coverage of any perf relevant functions, and maybe weekly, check today's master against 1 week ago, report any that have regressed by some % for review. 

@jackmott In that case https://github.com/PerfDotNet/BenchmarkDotNet/issues/155 looks like very promising.