@KevinRansom 2016-08-15
@jackmott If benchmark.net is a nuget package, then adding it will be great.