@jackmott 2016-08-15
@KevinRansom it is, but it pulls down a large set of dependencies I believe. Would you want it added to one of the existing projects?  Create a new project with it inside the scripts folder?